+++ main.sh
@@ -28,7 +28,7 @@ warn() {
     echo "::warning::$*"
 }
 info() {
-    echo >&2 "info: $*"
+    echo "info: $*"
 }
 
 if [[ $# -gt 0 ]]; then
@@ -59,7 +59,7 @@ for tool in "${tools[@]}"; do
                 # https://github.com/taiki-e/cargo-hack/releases
                 cargo-hack) latest_version="0.5.10" ;;
                 # https://github.com/taiki-e/cargo-llvm-cov/releases
-                cargo-llvm-cov) latest_version="0.1.13" ;;
+                cargo-llvm-cov) latest_version="0.1.15" ;;
                 # https://github.com/taiki-e/cargo-minimal-versions/releases
                 cargo-minimal-versions) latest_version="0.1.1" ;;
                 # https://github.com/taiki-e/parse-changelog/releases
@@ -144,5 +144,5 @@ for tool in "${tools[@]}"; do
         cargo-*) x cargo "${tool#cargo-}" --version ;;
         *) x "${tool}" --version ;;
     esac
-    echo >&2
+    echo
 done